part2A/requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
Flask-Limiter==3.5.0
Flask-OAuthlib==0.9.6
python-dotenv==1.0.0
authlib==1.2.1
